backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==5.0.9; python_version >= '3.6'
asgiref==3.4.1; python_version >= '3.6'
backports.zoneinfo==0.2.1; python_version < '3.9'
bidict==0.21.4; python_version >= '3.6'
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10; python_version >= '3'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.3; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-cors-headers==3.11.0
django==4.0.1
djangorestframework==3.13.1
idna==3.3; python_version >= '3'
kombu==5.2.3; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
prompt-toolkit==3.0.24; python_full_version >= '3.6.2'
psycopg2-binary==2.9.3
pyparsing==3.0.7; python_version >= '3.6'
python-engineio==4.3.1; python_version >= '3.6'
python-socketio[client]==5.5.1
pytz==2021.3
redis==4.1.1
requests==2.27.1
setuptools==59.6.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
